+++ app/classes/shared/main_controller.py
@@ -422,6 +422,7 @@ class Controller:
 
     def create_jar_server(
         self,
+        jar: str,
         server: str,
         version: str,
         name: str,
@@ -493,7 +494,7 @@ class Controller:
 
         # download the jar
         self.server_jars.download_jar(
-            server, version, os.path.join(server_dir, server_file), new_id
+            jar, server, version, os.path.join(server_dir, server_file), new_id
         )
 
         return new_id
@@ -932,6 +933,14 @@ class Controller:
                 self.servers.servers_list.pop(counter)
 
             counter += 1
+
+    def remove_unloaded_server(self, server_id):
+        try:
+            HelpersManagement.delete_scheduled_task_by_server(server_id)
+        except DoesNotExist:
+            logger.info("No scheduled jobs exist. Continuing.")
+        # remove the server from the DB
+        self.servers.remove_server(server_id)
 
     @staticmethod
     def clear_unexecuted_commands():